requirements.txt
pandas
numpy
requests
psycopg2-binary
python-dotenv
sqlalchemy
tensorflow
scikit-learn
